dialog on remove

+++ src/app/features/smokes-history/smokes-history.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Order } from '@datorama/akita';
 import { Observable, tap } from 'rxjs';
 import { SmokesQuery, today } from 'src/app/core/smokes/smokes.query';
@@ -9,15 +10,19 @@ import { ISmoke } from 'src/app/core/smokes/smokes.store';
   selector: 'app-smokes-history',
   templateUrl: './smokes-history.component.html',
   styleUrls: ['./smokes-history.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SmokesHistoryComponent {
-
-  smokes$: Observable<ISmoke[]> = this.smokes.selectAll({ sortBy: 'timestamp', sortByOrder: Order.DESC, filterBy: today });
+  smokes$: Observable<ISmoke[]> = this.smokes.selectAll({
+    sortBy: 'timestamp',
+    sortByOrder: Order.DESC,
+    filterBy: today,
+  });
   constructor(
     private smokes: SmokesQuery,
-    private service: SmokesService
-  ) { }
+    private service: SmokesService,
+    private dialog: MatDialog
+  ) {}
 
   trackById(index: number, smoke: ISmoke): string {
     return smoke.id;
@@ -28,8 +33,24 @@ export class SmokesHistoryComponent {
   }
 
   smokeRemoved(smoke: ISmoke) {
-    this.service.removeSmoke(smoke);
+    const ref = this.dialog.open(DialogComponent);
+    ref.afterClosed().subscribe((toBeRemoved) => {
+      if (toBeRemoved) {
+        this.service.removeSmoke(smoke);
+      }
+    });
   }
-
 }
 
+@Component({
+  template: `
+    <div class="container">
+      <span mat-dialog-title>Are you sure you want to remove that smoke?</span>
+      <div mat-dialog-actions>
+        <button mat-button [mat-dialog-close]="false">No</button>
+        <button mat-button [mat-dialog-close]="true" color="warn">Yes</button>
+      </div>
+    </div>
+  `,
+})
+export class DialogComponent {}
